Bot/requirements.txt
discord.py
requests
pymupdf